prototipo/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
pydantic[email]==2.7.1
python-dotenv==1.0.1
